Back to semantic versioning
I'd really prefer date-based; but pep 440 isn't clear to me how it resolves 'breaking' changes.

As a package with plugins; I'd imagine we would have 1.1, 1.2, etc in the event of adding plugins, and bumping to 2.0, 2.1, etc in the event that core updates (and most likely the plugins as well)

Actually, we might even need x.x.1 if we need to mark certain packages as bad in the dependencies.